First commit for the 2manysat's verifying compiler
This is a long overdue commit for the compiler. It includes the
following files:
Parser.hs -- Parser with compatibility for Int Expressions, while
		statements, assignment statements, and functions.
		Need to add Bool Expr and If statements.
Lexer.hs -- Lexer for the parser
Token.hs -- List of Tokens
Syntax.hs -- Data structures for the programs
Evaluate.hs -- Currently contains evaluation code for Int Expressions
only.
There is also a README, but it is blank so far.